Split website tests into multiple files
So I can run each of them separately, which is easier
